requirements-dev.txt
# Core web
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2

# ORM
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1

# Realtime
flask-socketio==5.3.6
python-socketio==5.11.3
python-engineio==4.9.1

# Security / auth
argon2-cffi==23.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
pyotp==2.9.0

# External services
redis==5.0.8
boto3==1.34.162
requests==2.32.3
psycopg2-binary==2.9.9

# HTML parsing (tests)
beautifulsoup4==4.12.3

# Testing
pytest==8.3.3
pytest-cov==5.0.0
celery==5.3.6
prometheus-client==0.20.0
